Map logical to physical addresses for file extents.
This is a first pass at implementing the missing functionality for #4.
For the simple one-device, non-RAID case, mapping logical to physical
addresses now works.
Initial reader code was added to read the data of the first non-inline
extent of a non-inline file. A follow-up change will implement full
extent support and sparse files.

Signed-off-by: Christian Blichmann <mail@blichmann.eu>